requirements.txt
python-shogi==1.0.15